explicit cast for now enum->float conversion, update settigs

+++ TESReloaded/Core/ShadowManager.cpp
@@ -849,7 +849,7 @@ bool ShadowManager::RenderShadowMaps() {
 	WasEnabledInt = ShadowsInteriors->Enabled;
 	TheRenderManager->SetupSceneCamera();
 	if (Player->GetWorldSpace() && ShadowsExteriors->Enabled) {
-		ShadowData->w = ShadowsExteriors->ShadowMode;	// Mode (0:off, 1:VSM, 2:ESM, 3: ESSM);
+		ShadowData->w = (UInt8)ShadowsExteriors->ShadowMode;	// Mode (0:off, 1:VSM, 2:ESM, 3: ESSM);
 		NiNode* PlayerNode = Player->GetNode();
 		D3DXVECTOR3 At;
